Fixed flinging mechanic
I fixed the flinging mechanic with a line of code that Brianna created for me. I have no idea what it does. I also made the empty gameobject instantiate bricks at the start of a touch and not spawn a brick before the start of a new touch. I also added the brick flying forward.